Separated out MySQL functions example into mysql.yml
