+++ composer.json
@@ -14,6 +14,7 @@
     "require-dev": {
         "doctrine/orm": "<2.2",
         "phpunit/phpunit": "~4.5",
+        "symfony/yaml": "~2.6",
         "zf1/zend-date": "~1.12",
         "zf1/zend-registry": "~1.12"
     },
